Mental-model/requirements.txt
faiss-cpu
groq
langchain-groq
pyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
flask
flask-cors